refactor(repositories): refactor project structure
