+++ repositories/user-repository.go
@@ -5,8 +5,8 @@ import (
 	"github.com/go-pg/pg"
 )
 
-// UserRepository interface
-type UserRepository interface {
+// IUserRepository interface
+type IUserRepository interface {
 	GetByID(id int) (*models.User, error)
 	Fetch() (*[]models.User, error)
 	Create(user *models.User) error
@@ -19,7 +19,7 @@ type userRepository struct {
 }
 
 // NewUserRepository returns a repository that implements the UserRepository interface
-func NewUserRepository(dbConn *pg.DB) UserRepository {
+func NewUserRepository(dbConn *pg.DB) IUserRepository {
 	return &userRepository{
 		DB: dbConn,
 	}
